appointment/requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.4
django-celery-results==2.0.1
django-cors-headers==3.7.0
djangorestframework==3.12.4
idna==2.10
kombu==5.1.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
coverage==5.5